Update imports to slack sdk v3
Before some legacy version of the slack sdk was used. This has now been
updated to use the right import for slack sdk v3.

Also the webclient is now instanciated only once on program start. This
has the nice side-effect that the program will exit with an error, when
the `SLACK_TOKEN` environment variable is not set.