refactor: rework event checking mehtod
There is no need to check for the current and the next day when we post
all events for the next day into the channel.
Also when the next day is Saturday we check the events for monday and
post them already on friday into the channel.
The refactoring also allows to have multiple events on one day and and
post all of them into the channel.